requirements.txt
django
djangorestframework
black
apscheduler
requests